test: consolidate 8 DB fixtures into shared MigratedDbFixture
Eight near-identical fixtures duplicated the container+migrate body.
Extract one base class; per-suite fixtures are now one-line subclasses.
