+++ tests/AHKFlowApp.Application.Tests/Hotstrings/HotstringDbFixture.cs
@@ -1,33 +1,9 @@
-using AHKFlowApp.Infrastructure.Persistence;
 using AHKFlowApp.TestUtilities.Fixtures;
-using Microsoft.EntityFrameworkCore;
 using Xunit;
 
 namespace AHKFlowApp.Application.Tests.Hotstrings;
 
-public sealed class HotstringDbFixture : IAsyncLifetime
-{
-    private readonly SqlContainerFixture _sql = new();
-
-    public string ConnectionString => _sql.ConnectionString;
-
-    public async Task InitializeAsync()
-    {
-        await _sql.InitializeAsync();
-        await using AppDbContext ctx = CreateContext();
-        await ctx.Database.MigrateAsync();
-    }
-
-    public Task DisposeAsync() => _sql.DisposeAsync();
-
-    public AppDbContext CreateContext()
-    {
-        DbContextOptions<AppDbContext> options = new DbContextOptionsBuilder<AppDbContext>()
-            .UseSqlServer(ConnectionString)
-            .Options;
-        return new AppDbContext(options);
-    }
-}
+public sealed class HotstringDbFixture : MigratedDbFixture;
 
 [CollectionDefinition("HotstringDb")]
 public sealed class HotstringDbCollection : ICollectionFixture<HotstringDbFixture>;
